requirements.txt
numpy
einops
scipy
pyyaml
tensorboardX
prettytable
timm
nuscenes-devkit
